Add InvolutionDiffeo and simplify NaturalToStandardDiffeo
